requirements.txt
transformers
pillow
torch
torchvision

# For remove BG
kornia
timm

easygui
